tests/unit/test_cli_assess_endpoint.py
import pytest
from typer.testing import CliRunner
from cli.main import app

runner = CliRunner()

def test_assess_endpoint_help():
    result = runner.invoke(app, ["assess-endpoint", "--help"])
    assert result.exit_code == 0
    assert "assess-endpoint" in result.stdout

# We will mock the actual script execution in the implementation
def test_assess_endpoint_mocked(mocker):
    # Mock the subprocess call or the internal function calling the script
    mocker.patch("cli.commands.assess_endpoint.run_assessment_script", return_value={"status": "ok"})
    
    result = runner.invoke(app, ["assess-endpoint", "--format", "json"])
    assert result.exit_code == 0
    # assert "ok" in result.stdout # Depends on how we print output
